requirements.txt
# Core libraries
pandas
faker
tqdm
pyyaml
neo4j
pytest

# Jupyter + plotting
jupyterlab
matplotlib

# Testing
pytest